+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Fix using Scrollspy component with default values
+- Fix Scrollspy if UIkit is loaded async
 
 ## 3.0.0 beta 39 (January 31, 2018)